requirements.txt
pyrogram
tgcrypto
ffmpeg-python